add Status to pending-requests et group-user
