+++ src/app/modules/group/components/member-list/member-list.component.ts
@@ -9,9 +9,9 @@ import { ActionFeedbackService } from 'src/app/shared/services/action-feedback.s
 import { Group } from '../../http-services/get-group-by-id.service';
 import { GetGroupChildrenService, GroupChild } from '../../http-services/get-group-children.service';
 import { GetGroupMembersService, Member } from '../../http-services/get-group-members.service';
-import { GroupUsersService } from '../../http-services/group-users.service';
+import { GroupUsersService, parseResults } from '../../http-services/group-users.service';
 import { Filter, GroupCompositionFilterComponent, TypeFilter } from '../group-composition-filter/group-composition-filter.component';
-import { displayResponseToast, parseResults } from './user-removal-response-handling';
+import { displayResponseToast } from './user-removal-response-handling';
 
 interface Column {
   sortable?: boolean,
